Text to Speech/requirements.txt
Flask~=2.2.3
python-dotenv~=1.0.0
requests~=2.28.2
elevenlabslib~=0.4.2